Fix AggregatorKernelAgentFactory naming
